Move files to appropreate folders; Add readme;
