+++ common/src/codegen/codegen-common/action-validator-registry.ts
@@ -1,26 +1,26 @@
-import { ITestStepRegular } from "rockmelonqa.common/file-defs/testCaseFile";
 import { ActionType } from "../../file-defs";
+import { ITestStepRegular } from "../../file-defs/testCaseFile";
 import { IActionTemplateParam } from "../types";
-import clear from "./action-validator-dotnet/clear";
-import click from "./action-validator-dotnet/click";
-import clickPopup from "./action-validator-dotnet/clickPopup";
-import closePopup from "./action-validator-dotnet/closePopup";
-import delay from "./action-validator-dotnet/delay";
-import gotoUrl from "./action-validator-dotnet/gotoUrl";
-import input from "./action-validator-dotnet/input";
-import inputByCode from "./action-validator-dotnet/inputByCode";
-import runCode from "./action-validator-dotnet/runCode";
-import selectOption from "./action-validator-dotnet/selectOption";
-import verifyAttribute from "./action-validator-dotnet/verifyAttribute";
-import verifyHasText from "./action-validator-dotnet/verifyHasText";
-import verifyHasValue from "./action-validator-dotnet/verifyHasValue";
-import verifyIsEditable from "./action-validator-dotnet/verifyIsEditable";
-import verifyIsHidden from "./action-validator-dotnet/verifyIsHidden";
-import verifyIsReadOnly from "./action-validator-dotnet/verifyIsReadOnly";
-import verifyIsVisible from "./action-validator-dotnet/verifyIsVisible";
-import verifyTitle from "./action-validator-dotnet/verifyTitle";
-import verifyTitleContains from "./action-validator-dotnet/verifyTitleContains";
-import verifyUrl from "./action-validator-dotnet/verifyUrl";
+import clear from "./action-validators/clear";
+import click from "./action-validators/click";
+import clickPopup from "./action-validators/clickPopup";
+import closePopup from "./action-validators/closePopup";
+import delay from "./action-validators/delay";
+import gotoUrl from "./action-validators/gotoUrl";
+import input from "./action-validators/input";
+import inputByCode from "./action-validators/inputByCode";
+import runCode from "./action-validators/runCode";
+import selectOption from "./action-validators/selectOption";
+import verifyAttribute from "./action-validators/verifyAttribute";
+import verifyHasText from "./action-validators/verifyHasText";
+import verifyHasValue from "./action-validators/verifyHasValue";
+import verifyIsEditable from "./action-validators/verifyIsEditable";
+import verifyIsHidden from "./action-validators/verifyIsHidden";
+import verifyIsReadOnly from "./action-validators/verifyIsReadOnly";
+import verifyIsVisible from "./action-validators/verifyIsVisible";
+import verifyTitle from "./action-validators/verifyTitle";
+import verifyTitleContains from "./action-validators/verifyTitleContains";
+import verifyUrl from "./action-validators/verifyUrl";
 
 /** A registry (map) of step validator for each ActionType */
 export const actionValidatorRegistryDotnet = new Map<ActionType, (step: ITestStepRegular) => string | undefined>();
