+++ common/src/codegen/codegen-common/action-validators/click.ts
@@ -1,7 +1,7 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequirePageAndElement } from "../action-validator-registry-dotnet";
+import { validateStepRequirePageAndElement } from "../action-validator-registry";
 
-/** Validate step with action Click Popup */
+/** Validate step with action Click */
 export default (step: ITestStepRegular) => {
   return validateStepRequirePageAndElement(step);
 };
