+++ common/src/codegen/codegen-common/action-validators/clickPopup.ts
@@ -1,0 +1,7 @@
+import { ITestStepRegular } from "../../../file-defs/testCaseFile";
+import { validateStepRequirePageAndElement } from "../action-validator-registry";
+
+/** Validate step with action Click Popup */
+export default (step: ITestStepRegular) => {
+  return validateStepRequirePageAndElement(step);
+};
