+++ common/src/codegen/codegen-common/action-validators/closePopup.ts
@@ -1,5 +1,3 @@
-import { ITestStep } from "rockmelonqa.common/file-defs/testCaseFile";
-import { getParameters } from "../../utils/stringUtils";
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
 
 /** Validate step with action Close Popup */
