+++ common/src/codegen/codegen-common/action-validators/delay.ts
@@ -1,5 +1,5 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequireData } from "../action-validator-registry-dotnet";
+import { validateStepRequireData } from "../action-validator-registry";
 
 /** Validate step with action Delay */
 export default (step: ITestStepRegular) => {
