+++ common/src/codegen/codegen-common/action-validators/input.ts
@@ -1,5 +1,5 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequirePageAndElement } from "../action-validator-registry-dotnet";
+import { validateStepRequirePageAndElement } from "../action-validator-registry";
 
 /** Validate step with action Input */
 export default (step: ITestStepRegular) => {
