+++ common/src/codegen/codegen-common/action-validators/runCode.ts
@@ -1,5 +1,5 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequireData } from "../action-validator-registry-dotnet";
+import { validateStepRequireData } from "../action-validator-registry";
 
 /** Validate step with action Run Code */
 export default (step: ITestStepRegular) => {
