+++ common/src/codegen/codegen-common/action-validators/selectOption.ts
@@ -1,7 +1,7 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequirePageAndElementAndData } from "../action-validator-registry-dotnet";
+import { validateStepRequirePageAndElementAndData } from "../action-validator-registry";
 
-/** Validate step with action Input By code */
+/** Validate step with action Clear */
 export default (step: ITestStepRegular) => {
   return validateStepRequirePageAndElementAndData(step);
 };
