+++ common/src/codegen/codegen-common/action-validators/verifyIsReadOnly.ts
@@ -1,7 +1,7 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequirePageAndElementAndData } from "../action-validator-registry-dotnet";
+import { validateStepRequirePageAndElement } from "../action-validator-registry";
 
 /** Validate step with action Clear */
 export default (step: ITestStepRegular) => {
-  return validateStepRequirePageAndElementAndData(step);
+  return validateStepRequirePageAndElement(step);
 };
