+++ common/src/codegen/codegen-common/action-validators/verifyTitleContains.ts
@@ -1,7 +1,7 @@
 import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequirePageAndElement } from "../action-validator-registry-dotnet";
+import { validateStepRequireData } from "../action-validator-registry";
 
 /** Validate step with action Clear */
 export default (step: ITestStepRegular) => {
-  return validateStepRequirePageAndElement(step);
+  return validateStepRequireData(step);
 };
