+++ common/src/codegen/codegen-common/readme.md
@@ -1,0 +1,3 @@
+# Codegen Common
+
+Contains common code for all CodeGen classes
