--- common/src/codegen/playwright-charp-common/action-validator-dotnet/verifyHasValue.ts
@@ -1,7 +1,0 @@
-import { ITestStepRegular } from "../../../file-defs/testCaseFile";
-import { validateStepRequirePageAndElementAndData } from "../action-validator-registry-dotnet";
-
-/** Validate step with action Clear */
-export default (step: ITestStepRegular) => {
-  return validateStepRequirePageAndElementAndData(step);
-};
